added more auth

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -7,6 +7,7 @@ use App\Models\Category;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -51,6 +52,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        $currenUserId = Auth::id();
+        if($currenUserId !== 1 ){
+            abort(403);
+        }
         return view('admin.categories.edit', compact('category'));
     }
 
@@ -74,6 +79,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $currenUserId = Auth::id();
+        if($currenUserId !== 1 ){
+            abort(403);
+        }
         $category->delete();
         return to_route('admin.categories.index')->with('message', "La Category '$category->name' è stata  eliminata");
     }
